Lots of BugFixes and Java Linting. Nodes can communicate with the SuperNode now

+++ assignment1/src/shared.thrift
@@ -1,5 +1,5 @@
 struct Machine {
        1: string ipAddress = "NULL",
-       2: i16 port,
-       3: i16 id,
+       2: i32 port,
+       3: i32 id,
 }
